% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univar_density_barplot_plotting.R
\name{univar_density_barplot_plotting}
\alias{univar_density_barplot_plotting}
\title{Title}
\usage{
univar_density_barplot_plotting(
  plotdata = data,
  x = NULL,
  group = NULL,
  grouporder = NULL,
  groupfill = NULL,
  grouplabel = NULL,
  plottitle = "",
  plot_xname = "Value",
  plot_yname = "Density",
  legendname = "Groups"
)
}
\arguments{
\item{legendname}{}
}
\description{
Title
}
